odroidw_meter.py
This application module utilizes the external 12-bit MCP3208 ADC connected to the Hardkernel Odroid W via SPI to be used as a real-time power meter. Current clamps in series with burden resistors are connected to the input channels of the MCP3208. The data is displayed in a TFT display and submitted via http to an external database.